pts: Modify upper bound check for file name in IMA event log
Make the upper bound check for the file name's length compatible with
LGTM's cpp/uncontrolled-allocation-size query, which causes an alert if
a user-controlled value is passed to malloc() without such a check.
However, it specifically expects such checks to have a quite specific
form, namely that the variable is on the left-hand side of a relational
operator (i.e. not ==).